pyproject.toml
[project]
name = "kule-liveblog-bot"
version = "0.1.0"
description = "A Discord bot to automatically transfer messages from Discord to Google Sheets for the Philippine Collegian's live blogs"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.7.1",
    "google-api-python-client>=2.196.0",
    "google-auth>=2.51.0",
    "python-dotenv>=1.2.2",
]

[tool.pyright]
reportUnknownMemberType = "none"
reportUnknownArgumentType = "none"

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]

[tool.ruff]
src = ["src"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "UP",
]